Rename files when they're finished, not when they're added
Downloading was taking a super long time. I'm not entirely sure why, but I think it has to do with the renaming and adding from a file or something, I'm not sure. I had to add a torrents collection to store all the torrents and what their names should be when the media is scanned.